[refactor] Add src/components/githubLogin - 깃헙 로그인 컴포넌트를 만들어 분리하였다.
